commit of kabob rule shell scripts
These scripts have been copied from the kabob project with minor
modifications.
The changes exist only in run-rules.sh (and pom.xml which is new) which
now uses the scripts/pom.xml file and makes use of the hanalyzer/rules
prefix when running a rule directory. There is no good way to access
the scripts in the kabob project so we have made explicit copies here.