updated compute-pms rule to use less memory
This revision includes additional constraints on the query that will
hopefully reduce its required memory. The rule only considers subsuming
parent nodes that are directly up the hierarchy and c1 and c2 are now
prohibited from being considered a parent (which should have been in
there in the first place). The rule name has also been updated to
include -go-bp-aael to differentiate it from the go-cc rule.